Gen-AI/Open-AI/requirements.txt
python-dotenv
openai
langchain
langchain_openai